load settings from newa config file
